UI/backend/pyproject.toml
# PEP 518/517 build system
[build-system]
requires = [
    "setuptools>=61.2.0",
    "wheel",
    "pyproject-hooks>=1.0.0",
]
build-backend = "setuptools.build_meta"

# Project metadata
[project]
name = "local_eval_backend"
description = "Local evaluation backend for the UI"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.7.0",
    "python-multipart>=0.0.5",
    "aiofiles>=23.1.0",
    "python-dotenv>=0.20.0",
    "pydantic>=1.10.0",
]

# Optional dependencies
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.16.0",
    "httpx>=0.23.0",
    "black>=22.3.0",
    "isort>=5.10.0",
    "mypy>=0.910",
]


[tool.setuptools]
package-dir = { "" = "app" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["app"]


